+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/callback/ScheduleBulkCaseForListingTest.java
@@ -2,7 +2,6 @@ package uk.gov.hmcts.reform.divorce.callback;
 
 import com.google.common.collect.ImmutableList;
 import org.apache.commons.lang3.tuple.Pair;
-import org.junit.Ignore;
 import org.junit.Test;
 import uk.gov.hmcts.reform.divorce.model.ccd.CaseLink;
 import uk.gov.hmcts.reform.divorce.model.ccd.CollectionMember;
@@ -32,7 +31,6 @@ public class ScheduleBulkCaseForListingTest extends CcdSubmissionSupport {
     private static final int  POOL_INTERVAL_IN_MILLIS = 1000;
 
     @Test
-    @Ignore
     public void whenScheduleBulkCaseForListing_thenIndividualCasesShouldBeUpdated() throws Exception {
         final UserDetails user1 = createCitizenUser();
         final UserDetails user2 = createCitizenUser();
